+++ library/src/main/java/com/pengrad/telegrambot/request/business/RepostStory.kt
@@ -9,13 +9,13 @@ import com.pengrad.telegrambot.utility.kotlin.requestParameter
 class RepostStory private constructor(
     businessConnectionId: String,
     fromChatId: Long,
-    fromStoryId: Int,
+    fromStoryId: Long,
     activePeriod: Int,
     postToChatPage: Boolean?,
     protectContent: Boolean?
 ): KBaseRequest<RepostStory, RepostStoryResponse>(RepostStoryResponse::class) {
 
-    constructor(businessConnectionId: String, fromChatId: Long, fromStoryId: Int, activePeriod: Int) : this(
+    constructor(businessConnectionId: String, fromChatId: Long, fromStoryId: Long, activePeriod: Int) : this(
         businessConnectionId = businessConnectionId,
         fromChatId = fromChatId,
         fromStoryId = fromStoryId,
@@ -26,7 +26,7 @@ class RepostStory private constructor(
 
     val businessConnectionId: String by requestParameter(businessConnectionId, customParameterName = "business_connection_id")
     val fromChatId: Long by requestParameter(fromChatId, customParameterName = "from_chat_id")
-    val fromStoryId: Int by requestParameter(fromStoryId, customParameterName = "from_story_id")
+    val fromStoryId: Long by requestParameter(fromStoryId, customParameterName = "from_story_id")
     val activePeriod: Int by requestParameter(activePeriod, customParameterName = "active_period")
 
     var postToChatPage: Boolean? by optionalRequestParameter(postToChatPage, customParameterName = "post_to_chat_page")
